refactor: move scheduler to feature folder
